Refactor Public Key to Private Key
I noticed is that this method and naming is a misnomer and should probably be updated before going stable. SSH public keys aren't used to connect, private keys are. The Public key is setup on the server as the "lock" in authorized_keys and the private key is, well, the key to that lock.

Citation: https://explainshell.com/explain?cmd=ssh+-i